Address PR #869 review: move utils and merge helpers, break import cycles
Move annotation_utils into TraceUtils, keep merge_capture_trace_into_graph in
Trace2Tree, load traces without TreePerfAnalyzer, and document lazy cross-package
imports in import-linter so CI stays acyclic at import time.
