+++ .importlinter
@@ -8,3 +8,6 @@ type = acyclic_siblings
 ancestors =
     TraceLens
 depth = 10
+ignore_imports =
+    TraceLens.Trace2Tree.trace_capture_merge_experimental -> TraceLens.TraceUtils.annotation_utils
+    TraceLens.TraceUtils.split_inference_trace_annotation -> TraceLens.Trace2Tree.inference_iteration_roots
