+++ tests/test_split_inference_trace_annotation.py
@@ -18,7 +18,7 @@ No trace files are written; everything operates on in-memory dicts.
 """
 
 from TraceLens.TraceUtils import split_inference_trace_annotation as split
-from TraceLens.annotation_utils import IterationAnnotation
+from TraceLens.TraceUtils.annotation_utils import IterationAnnotation
 
 # --------------------------------------------------------------------------- #
 # Dummy-trace builder
